+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [unreleased]
+
+### Fixed
+
+* hasura: Fixed metadata generation.
+
 ## [5.2.1] - 2022-07-02
 
 ### Fixed